use argh from_str_fn feature
